# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enerji Yönetim Sistemi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /swot:
    get:
      operationId: listSwotItems
      tags: [swot]
      summary: SWOT maddeleri listesi
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SwotItem"
    post:
      operationId: createSwotItem
      tags: [swot]
      summary: SWOT maddesi ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwotItem"
  /swot/{id}:
    patch:
      operationId: updateSwotItem
      tags: [swot]
      summary: SWOT maddesi güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwotUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwotItem"
    delete:
      operationId: deleteSwotItem
      tags: [swot]
      summary: SWOT maddesi sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── TARGETS ───────────────────────────────────────────────
...
  /seu:
    get:
      operationId: listSeu
      tags: [seu]
      summary: Önemli enerji kullanımları listesi
      parameters:
        - name: unitId
          in: query
          schema:
            type: integer
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SeuItem"
    post:
      operationId: createSeu
      tags: [seu]
      summary: ÖEK ekle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeuInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeuItem"
  /seu/{id}:
    patch:
      operationId: updateSeu
      tags: [seu]
      summary: ÖEK güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeuUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeuItem"
    delete:
      operationId: deleteSeu
      tags: [seu]
      summary: ÖEK sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DASHBOARD ─────────────────────────────────────────────
...
components:
  schemas:
...
    SwotItem:
      type: object
      required: [id, category, title, score, impact]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        category:
          type: string
          description: "strengths | weaknesses | opportunities | threats"
        title:
          type: string
        description:
          type: ["string", "null"]
        score:
          type: integer
          description: "1-5 puan"
        impact:
          type: string
          description: "yuksek | orta | dusuk"
        createdAt:
          type: string
    SwotInput:
      type: object
      required: [category, title, score, impact]
      properties:
        unitId:
          type: integer
        category:
          type: string
        title:
          type: string
        description:
          type: string
        score:
          type: integer
        impact:
          type: string
    SwotUpdate:
      type: object
      properties:
        unitId:
          type: integer
        category:
          type: string
        title:
          type: string
        description:
          type: string
        score:
          type: integer
        impact:
          type: string

    # ── Risk ─────────────────────────────────────────────────
...
    SeuItem:
      type: object
      required: [id, name, category, annualKwh, percentage, priority]
      properties:
        id:
          type: integer
        unitId:
          type: ["integer", "null"]
        name:
          type: string
        category:
          type: string
        annualKwh:
          type: number
        percentage:
          type: number
        priority:
          type: integer
        targetReductionPercent:
          type: ["number", "null"]
        responsible:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    SeuInput:
      type: object
      required: [name, category]
      properties:
        unitId:
          type: integer
        name:
          type: string
        category:
          type: string
        annualKwh:
          type: number
        percentage:
          type: number
        priority:
          type: integer
        targetReductionPercent:
          type: number
        responsible:
          type: string
        notes:
          type: string
    SeuUpdate:
      type: object
      properties:
        unitId:
          type: integer
        name:
          type: string
        category:
          type: string
        annualKwh:
          type: number
        percentage:
          type: number
        priority:
          type: integer
        targetReductionPercent:
          type: number
        responsible:
          type: string
        notes:
          type: string

    # ── Dashboard ─────────────────────────────────────────────
...